2020.11.04_1

+++ entrypoint.sh
@@ -5,6 +5,9 @@ echo HOST_APP_DIR: ${HOST_APP_DIR}
 set -e
 
 # Remove a potentially pre-existing server.pid for Rails.
-rm -f $APP_DIR/tmp/pids/server.pid
+if test -f "$APP_DIR/tmp/pids/server.pid"; then
+    rm -f $APP_DIR/tmp/pids/server.pid
+fi
+# rm -f $APP_DIR/tmp/pids/server.pid
 # Then exec the container's main process (what's set as CMD in the Dockerfile).
 exec "$@"
